V2.16.4

+++ backend-django/app/eric/core/db/migrations.py
@@ -1,9 +1,8 @@
-# -*- coding: utf-8 -*-
 #
-# Copyright (C) 2016-2020 TU Muenchen and contributors of ANEXIA Internetdienstleistungs GmbH
+# Copyright (C) 2016-present TU Muenchen and contributors of ANEXIA Internetdienstleistungs GmbH
 # SPDX-License-Identifier: AGPL-3.0-or-later
 #
-from __future__ import unicode_literals
+
 from django.core.management.sql import emit_post_migrate_signal
 from django.db import migrations
 
@@ -16,10 +15,10 @@ class PermissionContentTypeApplyMigration(migrations.Migration):
         db_alias = schema_editor.connection.alias
         emit_post_migrate_signal(2, False, db_alias)
 
-        return super(PermissionContentTypeApplyMigration, self).apply(project_state, schema_editor, collect_sql)
+        return super().apply(project_state, schema_editor, collect_sql)
 
     def unapply(self, project_state, schema_editor, collect_sql=False):
-        ret = super(PermissionContentTypeApplyMigration, self).unapply(project_state, schema_editor, collect_sql)
+        ret = super().unapply(project_state, schema_editor, collect_sql)
         # we need current content types and migrations
         # therefore we need to emit a post migrate signal BEFORE we access those
         # https://code.djangoproject.com/ticket/23422
@@ -32,7 +31,7 @@ class PermissionContentTypeApplyMigration(migrations.Migration):
 class PermissionMigrationHelper:
     @staticmethod
     def convert_string_permissions_list_to_model_permissions(apps, schema_editor, string_permission_list):
-        """ Converts a given string permission list to a set of Django permissions """
+        """Converts a given string permission list to a set of Django permissions"""
         ContentType = apps.get_model("contenttypes", "ContentType")
         Permission = apps.get_model("auth", "Permission")
 
@@ -43,29 +42,32 @@ class PermissionMigrationHelper:
         for string_permission in string_permission_list:
             # string permission looks as follows:
             # ['codename', 'app', 'Model']
-            assert len(string_permission) == 3, \
-                "Parameter string_permission_list is an array and must contain arrays of exactly 3 elements"
+            assert (
+                len(string_permission) == 3
+            ), "Parameter string_permission_list is an array and must contain arrays of exactly 3 elements"
 
             # get content type of the model
             if isinstance(string_permission[1], list):
-                content_type = ContentType.objects.using(db_alias).filter(
-                    app_label__in=string_permission[1],
-                    model=string_permission[2]
-                ).first()
+                content_type = (
+                    ContentType.objects.using(db_alias)
+                    .filter(app_label__in=string_permission[1], model=string_permission[2])
+                    .first()
+                )
             else:
                 content_type = ContentType.objects.using(db_alias).get(
-                    app_label=string_permission[1],
-                    model=string_permission[2]
+                    app_label=string_permission[1], model=string_permission[2]
                 )
 
             # get the permission via code name and content type
             permissions = Permission.objects.using(db_alias).filter(
-                codename=string_permission[0],
-                content_type=content_type
+                codename=string_permission[0], content_type=content_type
             )
-            assert len(permissions) <= 1, \
-                "There are %(number_permission)d (expected 1) permissions with the name %(permission_name)s" % \
-                {'number_permission': len(permissions), 'permission_name': string_permission[0]}
+            assert (
+                len(permissions) <= 1
+            ), "There are %(number_permission)d (expected 1) permissions with the name %(permission_name)s" % {
+                "number_permission": len(permissions),
+                "permission_name": string_permission[0],
+            }
 
             if len(permissions) == 1:
                 # get the first (and only permission)
@@ -79,7 +81,7 @@ class PermissionMigrationHelper:
 
     @staticmethod
     def add_permissions_to_role(apps, schema_editor, permissions, role):
-        """ adds a list of django permissions to a role """
+        """adds a list of django permissions to a role"""
         RolePermissionAssignment = apps.get_model("projects", "RolePermissionAssignment")
         db_alias = schema_editor.connection.alias
 
@@ -88,27 +90,23 @@ class PermissionMigrationHelper:
 
         for permission in permissions:
             if not RolePermissionAssignment.objects.using(db_alias).filter(role=role, permission=permission).exists():
-                role_permission_assignments.append(
-                    RolePermissionAssignment(role=role, permission=permission)
-                )
+                role_permission_assignments.append(RolePermissionAssignment(role=role, permission=permission))
 
         # bulk create role permission assignments
         RolePermissionAssignment.objects.using(db_alias).bulk_create(role_permission_assignments)
 
     @staticmethod
     def remove_permission_from_role(apps, schema_editor, permissions, role):
-        """ removes a list of django permissions from a role """
+        """removes a list of django permissions from a role"""
         RolePermissionAssignment = apps.get_model("projects", "RolePermissionAssignment")
         db_alias = schema_editor.connection.alias
 
         for permission in permissions:
-            RolePermissionAssignment.objects.using(db_alias).filter(
-                role=role, permission=permission
-            ).delete()
+            RolePermissionAssignment.objects.using(db_alias).filter(role=role, permission=permission).delete()
 
     @staticmethod
     def forwards_func_project_manager(permissions_to_add, apps, schema_editor):
-        """ Migration forward func for adding permissions to the project manager """
+        """Migration forward func for adding permissions to the project manager"""
         Role = apps.get_model("projects", "Role")
 
         db_alias = schema_editor.connection.alias
@@ -117,18 +115,13 @@ class PermissionMigrationHelper:
             apps, schema_editor, permissions_to_add
         )
 
-        project_manager_role = Role.objects.using(db_alias).filter(
-            default_role_on_project_create=True
-        ).first()
+        project_manager_role = Role.objects.using(db_alias).filter(default_role_on_project_create=True).first()
 
-        PermissionMigrationHelper.add_permissions_to_role(
-            apps, schema_editor,
-            permissions, project_manager_role
-        )
+        PermissionMigrationHelper.add_permissions_to_role(apps, schema_editor, permissions, project_manager_role)
 
     @staticmethod
     def reverse_func_project_manager(permissions_to_add, apps, schema_editor):
-        """ Migration reverse func for removing permissions to the project manager """
+        """Migration reverse func for removing permissions to the project manager"""
         Role = apps.get_model("projects", "Role")
 
         db_alias = schema_editor.connection.alias
@@ -137,18 +130,13 @@ class PermissionMigrationHelper:
             apps, schema_editor, permissions_to_add
         )
 
-        project_manager_role = Role.objects.using(db_alias).filter(
-            default_role_on_project_create=True
-        ).first()
+        project_manager_role = Role.objects.using(db_alias).filter(default_role_on_project_create=True).first()
 
-        PermissionMigrationHelper.remove_permission_from_role(
-            apps, schema_editor,
-            permissions, project_manager_role
-        )
+        PermissionMigrationHelper.remove_permission_from_role(apps, schema_editor, permissions, project_manager_role)
 
     @staticmethod
     def forwards_func_observer(permissions_to_add, apps, schema_editor):
-        """ Migration forward func for adding permissions to the project manager """
+        """Migration forward func for adding permissions to the project manager"""
         Role = apps.get_model("projects", "Role")
 
         db_alias = schema_editor.connection.alias
@@ -157,18 +145,13 @@ class PermissionMigrationHelper:
             apps, schema_editor, permissions_to_add
         )
 
-        observer_role = Role.objects.using(db_alias).filter(
-            name="Observer"
-        ).first()
+        observer_role = Role.objects.using(db_alias).filter(name="Observer").first()
 
-        PermissionMigrationHelper.add_permissions_to_role(
-            apps, schema_editor,
-            permissions, observer_role
-        )
+        PermissionMigrationHelper.add_permissions_to_role(apps, schema_editor, permissions, observer_role)
 
     @staticmethod
     def reverse_func_observer(permissions_to_add, apps, schema_editor):
-        """ Migration reverse func for removing permissions to the project manager """
+        """Migration reverse func for removing permissions to the project manager"""
         Role = apps.get_model("projects", "Role")
 
         db_alias = schema_editor.connection.alias
@@ -177,14 +160,9 @@ class PermissionMigrationHelper:
             apps, schema_editor, permissions_to_add
         )
 
-        observer_role = Role.objects.using(db_alias).filter(
-            name="Observer"
-        ).first()
+        observer_role = Role.objects.using(db_alias).filter(name="Observer").first()
 
-        PermissionMigrationHelper.remove_permission_from_role(
-            apps, schema_editor,
-            permissions, observer_role
-        )
+        PermissionMigrationHelper.remove_permission_from_role(apps, schema_editor, permissions, observer_role)
 
 
 class ContentTypeMigrationHelper:
